pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-startup-simulator"
version = "1.0.0"
description = "AI Startup Founder Simulator — an OpenEnv environment for training and evaluating AI agents on real-world business strategy decisions."
authors = [
    { name = "Rajiv Ranjan Jha" }
]
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
    "gradio>=5.12.0",
    "fastapi",
    "uvicorn",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.1.0",
    "shimmy>=1.3.0",
    "pillow",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "websockets",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["server*"]
